crew-youtube/pyproject.toml
[tool.poetry]
name="automate-youtube-with-crewai"
version="0.1.0"
description="CrewAI"
authors=["Gen AI"]

[tool.poetry.dependencies]
python=">=3.10.0,<3.12"
crewai={extras =["tools"], version ="^0.19.0"}
pydantic="^2.6.3"
load-dotenv="^0.1.0"
crewai-tools="^0.0.15"

[tool.pyright]
#https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes=true
exclude=[".cache"]

[tool.ruff]
#https://docs.astral.sh/ruff/configuration/#config-file-discovery
select =['E','W','F','I','B','C4','ARG','SIM']
ignore=['W291','W292','W293']

[build-system]
requires=["poetry-core>=1.1.0"]
build-backend ="poetry.core.masonry.api"